Separately profile the time taken waiting for bytes to arrive and the time taken waiting for bytes to be sent; Removed unused profile variable in cloud_comm.c;

+++ sdk/cloud_comm/src/protocols/ott_protocol/ott_protocol.c
@@ -86,7 +86,7 @@ static void *ott_calloc(size_t num, size_t size)
 #endif
 
 #ifdef PROTO_TIME_PROFILE
-static uint32_t proto_begin;
+static uint64_t proto_begin;
 #endif
 
 static uint32_t current_polling_interval = INIT_POLLING_MS;
